requirements.txt
asyncio>=3.4.3
qrcode>=6.1
praw>=7.4.0
discord.py>=1.7.0
discord-components>=2.0.1
wavelink==1.0.0b30
python-whois >=0.7.3
statcord.py>=3.1.0